added recommended usage to settings (#66)
The Yearly Review plugin creates a topic automatically on the first day of ever year containing a summary of forum activity. It is highly recommended that site owners set it up to post it to the staff category or another private category, so they can view it first and optionally edit it. They might like to add a celebratory note!