FIX: Fixes topic creation double-ups (#69)
Followup to 98012d8c923c9c9ad0028e3c8c44fffc13556ea6,
if the first review topic for the year is deleted, then
that was permanently affecting this check and making it
incorrect, leading to the yearly review topic being created
as many times as the sidekiq job was run. This is because
the find_by method was finding the first topic custom field,
which would always have a deleted topic and always be null.

Now we filter the check and make it so it doesn't include
deleted topics.